開催日時の表示を変更
