--- src/app/project/template/templateTable.ts
@@ -1,24 +1,0 @@
-export const timeScheduleDay1 = {
-  column: ["時間", "団体名", "内容"],
-  data: [
-    ["10:15〜10:45", "名古屋工業大学吹奏楽団", "吹奏楽演奏"],
-    ["10:45〜10:47", "転換"],
-    ["10:47〜11:17", "Prego", "アカペラ"],
-    ["11:17〜11:19", "転換"],
-    ["11:19〜11:49", "名城大学薬学部奇術部 ROOK", "ジャグリング"],
-    ["11:49〜11:51", "転換"],
-    ["11:51〜12:11", "金城学院大学ダンス部 BEAT", "ダンス"],
-  ],
-  widthWeight: [2, 5, 3],
-};
-
-export const timeScheduleDay2 = {
-  column: ["時間", "団体名"],
-  data: [
-    ["15:15〜15:30", "愛知flavor"],
-    ["15:30〜15:40", "Paradox Risk"],
-    ["15:40〜15:50", "WONDER SNAKE"],
-    ["15:55〜16:10", "幻"],
-  ],
-  widthWeight: [1, 2],
-};
