+++ src/components/Project/ProjectContent/ProjectContent.tsx
@@ -7,6 +7,10 @@ interface ProjectContentProps {
 }
 
 export default function ProjectContent({ projectData }: ProjectContentProps) {
+  let isDateAlreadyGenerated = false;
+  // 開催日時
+  const date = ["11/16(土)", "11/17(日)"];
+
   return (
     <>
       {Object.keys(projectData).map((key) => {
@@ -15,6 +19,27 @@ export default function ProjectContent({ projectData }: ProjectContentProps) {
         // 非表示項目の場合
         if (data?.isHidden ?? false) {
           return null;
+        }
+
+        // 例外項目
+        if (!isDateAlreadyGenerated && (key === "day1" || key === "day2")) {
+          isDateAlreadyGenerated = true;
+          return (
+            <ContentBox key={key} title="開催日時">
+              {projectData?.day1 && (
+                <>
+                  1日目{`(${date[0]})`}: {projectData.day1.content}
+                  <br />
+                </>
+              )}
+              {projectData?.day2 && (
+                <>
+                  2日目{`(${date[1]})`}: {projectData.day2.content}
+                  <br />
+                </>
+              )}
+            </ContentBox>
+          );
         }
 
         // contentが配列の場合
